pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "marderlab-tools"
version = "0.1.0"
description = "Production analysis pipelines for Marder Lab electrophysiology data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
analysis = [
  "matplotlib>=3.8",
  "pyabf>=2.3",
  "scipy>=1.10",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
marder = "marderlab_tools.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
